# Conflicts:
#	app/src/debug/java/com/gat/data/DebugLocalMessageDataSource.java
#	app/src/debug/java/com/gat/data/DebugMessageDataSource.java
#	app/src/main/java/com/gat/app/screen/ParcelableScreen.java
#	app/src/main/java/com/gat/app/screen/ScreenPresenterFactory.java
#	app/src/main/java/com/gat/common/util/CommonUtil.java
#	app/src/main/java/com/gat/data/firebase/FirebaseService.java
#	app/src/main/java/com/gat/data/firebase/FirebaseServiceImpl.java
#	app/src/main/java/com/gat/data/user/PaperUserDataSource.java
#	app/src/main/java/com/gat/dependency/PresenterComponent.java
#	app/src/main/java/com/gat/dependency/PresenterModule.java
#	app/src/main/java/com/gat/domain/UseCaseFactory.java
#	app/src/main/java/com/gat/domain/impl/UseCaseFactoryImpl.java
#	app/src/main/java/com/gat/feature/login/LoginActivity.java
#	app/src/main/java/com/gat/feature/message/MessageActivity.java
#	app/src/main/java/com/gat/feature/message/item/MessageItemBuilder.java
#	app/src/main/java/com/gat/feature/message/presenter/GroupMessagePresenterImpl.java
#	app/src/main/java/com/gat/feature/message/presenter/MessagePresenterImpl.java
#	app/src/main/java/com/gat/feature/message/viewholder/MessageViewHolder.java
#	app/src/main/java/com/gat/feature/register/RegisterActivity.java
#	app/src/main/java/com/gat/feature/start/StartActivity.java
#	app/src/main/java/com/gat/repository/MessageRepository.java
#	app/src/main/java/com/gat/repository/UserRepository.java
#	app/src/main/java/com/gat/repository/datasource/MessageDataSource.java
#	app/src/main/java/com/gat/repository/impl/MessageRepositoryImpl.java

+++ app/src/main/java/com/gat/repository/entity/book/BookDetailEntity.java
@@ -180,7 +180,7 @@ public class BookDetailEntity {
         private String author = Strings.EMPTY;
 
         @SerializedName("rateAvg")
-        private int rateAvg;
+        private double rateAvg;
 
         @SerializedName("rateCount")
         private int rateCount;
@@ -208,7 +208,7 @@ public class BookDetailEntity {
             return author;
         }
 
-        public int getRateAvg() {
+        public double getRateAvg() {
             return rateAvg;
         }
